Core:  - Fixed serialization of references.  - Objects are now reference-counted.
Engine:
 - Removed shared_ptr from the code.

Editor:
 - Added startup priority to plugins (fixes toolbar items ordering).
 - Added extra toolbar for mode-specific options.
 - Started adding additive selection mode.
 - Added custom cursors for additive selections.

+++ src/Editor/PluginManagerFrame.cpp
@@ -31,7 +31,7 @@ void PluginManagerFrame::InitControl()
 	const std::vector<Plugin*>& plugs = plugins->getPlugins();
 
 	// Populate the list with the plugins.
-	for( uint i = 0; i < plugs.size(); i++ )
+	for( size_t i = 0; i < plugs.size(); i++ )
 	{
 		Plugin* plugin = plugs[i];
 		PluginMetadata metadata = plugin->getMetadata();
